#41 PassiveCheatkeyAdapter 와 PassiveCheatkeyData 분리
