+++ jibcon/src/main/java/com/sm_arts/jibcon/app/cheatkey/PassiveCheatkeyData.java
@@ -1,0 +1,17 @@
+package com.sm_arts.jibcon.app.cheatkey;
+
+/**
+ * Created by woojinkim on 2017. 7. 17..
+ */
+
+public class PassiveCheatkeyData{
+    public int imgBtnID;
+    public String inputTv;
+    public String outputTv;
+
+    public PassiveCheatkeyData(int imgBtnID, String inputTv, String outputTv){
+        this.imgBtnID = imgBtnID;
+        this.inputTv = inputTv;
+        this.outputTv = outputTv;
+    }
+}
